refactor/#4: 라우터 구조 수정
