+++ sample-community/app/_layout.tsx
@@ -1,9 +1,11 @@
-import { Stack } from "expo-router";
+import { Stack } from 'expo-router';
 
 export default function RootLayout() {
-  return <Stack
-    screenOptions={{
-      headerTitle: "",
-    }}
-  />;
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+      }}
+    />
+  );
 }
